requirements.txt
fastapi
uvicorn
firebase-admin
httpx
python-dotenv
pydantic
aiofiles
